requirements.txt
tcod>=11.15
numpy>=1.18
imageio>=2.36
pillow
pygame
fastapi>=0.104.0
uvicorn>=0.24.0